core/client-kafka/pyproject.toml
[tool.poetry]
name = "client-kafka"
version = "0.1.0"
description = ""
authors = ["Matus-Abrhan <xabrhan@stuba.sk>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
kafka-python-ng = "^2.2.2"
winevt = "^0.0.11"
xmltodict = "^0.13.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
client-kafka = "client_kafka.run:run"
